Update init.d support
init.d support has not been touched by recent features, see e.g.
commit e78bf3196207. No idea how many distro still uses it, but in order
to avoid having a piece of code rotting on its own, it would be wise to
leverage `ethercatctl`, so any changes will be automatically exposed in
init.d too.

This is exactly what this patch tries to achieve. To be able to retain
backward compatibility, the following steps have been performed:

- add `-c` to `ethercatctl` to (optionally) load custom configuration:
  init.d uses `$ETC/sysconfig/ethercat` instead of `$ETC/ethercat.conf`
- share `ethercat.conf` code between systemd and init.d
- add `--disable-initd` to configure to be able to entirely remove
  init.d support (by default init.d support is enabled)

I do not have a system to phisycally try the new init.d script, but the
implementation should be quite straightforward.

AFAIK the only breaking change is that the `/etc/init.d/ethercat status`
command now does not call any rc_... function if `/etc/rc.status` is
found. Not sure what that means but properly implementing that would
basically require having to duplicate of the status code, partially
defeating the purpose of this patch.